feat: separate types separately into types package

+++ scripts/build.js
@@ -18,21 +18,22 @@ async function buildAll(targets) {
 }
 
 async function build(target) {
-  const pkgDir = path.resolve(`packages/${target}`)
+  if (target !== 'types') {
+    const pkgDir = path.resolve(`packages/${target}`)
 
-  await fs.remove(`${pkgDir}/dist`)
+    await fs.remove(`${pkgDir}/dist`)
 
-  const env = 'production'
-
-  await execa(
-    'rollup',
-    [
-      '-c',
-      '--environment',
-      [`NODE_ENV:${env}`, `TARGET:${target}`, formats ? `FORMATS:${formats}` : ``].join(',')
-    ],
-    {
-      stdio: 'inherit'
-    }
-  )
+    const env = 'production'
+    await execa(
+      'rollup',
+      [
+        '-c',
+        '--environment',
+        [`NODE_ENV:${env}`, `TARGET:${target}`, formats ? `FORMATS:${formats}` : ``].join(',')
+      ],
+      {
+        stdio: 'inherit'
+      }
+    )
+  }
 }
